[refactor] access토큰으로 친구목록 불러오기 서비스 위치변경(KakaoFunding#161)
